refactor: move Colors class to common utilities (#2601)
Move Colors class from migrate_notes_to_github.py to scripts/common.py
for reusability across multiple scripts.

Changes:
- Add Colors class to scripts/common.py with improved docstrings
- Add disable() method for non-TTY output
- Update import in migrate_notes_to_github.py
- Remove duplicate Colors class definition

Closes #2601
